Move search page to index page
